fix(resolutions): Remove useless sleep call
There was a case where resizing the nav bar to something too small to
contain it caused the unrendered parts of the nav bar to appear
white. At first I thought it was because I took the screenshot too
fast, so I tried sleeping for a little while before taking
screenshots. Later, I realized what was actually going on.