fix(screenshot): Force name to be attempted resolution, not actual
For screenshots that resize the browser window to dimensions smaller than what is
supported in Selenium, this tool will resize the window to the closest value it can,
and then name the screenshot based on what the resolution *was* at the time when it
was taken. Instead, this fix causes the desired resolution to always appear in the name.

This is useful because as newer versions of selenium are released, the smallest available
resolution changes with it. This gives a screenshot a new name, and registers as a brand
new screenshot, instead of an overwritten screenshot that is slightly more narrow than before.